fix: resolve Float/Double type mismatch in map bounds calculation
- Changed boundsMinX/Y/MaxX/Y from Double? to Float?
- Updated MqttService.publishMap() signature to accept Float?
- Fixes Kotlin compile error on location.x/y (Float? vs Double?)

+++ app/src/main/java/com/example/temicontroller/MainActivity.kt
@@ -728,10 +728,10 @@ class MainActivity : AppCompatActivity() {
                 
                 // Compute world coordinate bounds from Floor locations
                 val currentFloor = r.getCurrentFloor()
-                var boundsMinX: Double? = null
-                var boundsMinY: Double? = null
-                var boundsMaxX: Double? = null
-                var boundsMaxY: Double? = null
+                var boundsMinX: Float? = null
+                var boundsMinY: Float? = null
+                var boundsMaxX: Float? = null
+                var boundsMaxY: Float? = null
                 if (currentFloor != null && currentFloor.locations.isNotEmpty()) {
                     val xs = currentFloor.locations.map { it.x }
                     val ys = currentFloor.locations.map { it.y }
